requirement.txt
opencv-python
Pillow
pytesseract
spacy